HERCULES_0.0.1

+++ sgi-csp-service/src/main/java/org/crue/hercules/sgi/csp/service/impl/ProyectoEntidadFinanciadoraServiceImpl.java
@@ -180,7 +180,7 @@ public class ProyectoEntidadFinanciadoraServiceImpl implements ProyectoEntidadFi
   private void validateData(ProyectoEntidadFinanciadora proyectoEntidadFinanciadora) {
     Assert.isTrue(
         proyectoEntidadFinanciadora.getPorcentajeFinanciacion() == null
-            || proyectoEntidadFinanciadora.getPorcentajeFinanciacion() >= 0,
+            || proyectoEntidadFinanciadora.getPorcentajeFinanciacion().floatValue() >= 0,
         "PorcentajeFinanciacion no puede ser negativo");
 
     if (proyectoEntidadFinanciadora.getFuenteFinanciacion() != null) {
