+++ sgi-csp-service/src/test/java/org/crue/hercules/sgi/csp/integration/ProyectoEntidadFinanciadoraIT.java
@@ -1,5 +1,6 @@
 package org.crue.hercules.sgi.csp.integration;
 
+import java.math.BigDecimal;
 import java.util.Collections;
 
 import org.assertj.core.api.Assertions;
@@ -193,8 +194,8 @@ public class ProyectoEntidadFinanciadoraIT extends BaseIT {
         .as("getFuenteFinanciacion().getId()").isEqualTo(1L);
     Assertions.assertThat(proyectoEntidadFinanciadora.getTipoFinanciacion().getId()).as("getTipoFinanciacion().getId()")
         .isEqualTo(1L);
-    Assertions.assertThat(proyectoEntidadFinanciadora.getPorcentajeFinanciacion()).as("getPorcentajeFinanciacion()")
-        .isEqualTo(20);
+    Assertions.assertThat(proyectoEntidadFinanciadora.getPorcentajeFinanciacion().floatValue())
+        .as("getPorcentajeFinanciacion()").isEqualTo(20F);
     Assertions.assertThat(proyectoEntidadFinanciadora.getAjena()).as("getAjena()").isEqualTo(false);
   }
 
@@ -227,7 +228,7 @@ public class ProyectoEntidadFinanciadoraIT extends BaseIT {
       .entidadRef("entidad-" + String.format("%03d", id == null ? 0 : id))
       .fuenteFinanciacion(fuenteFinanciacion)
       .tipoFinanciacion(tipoFinanciacion)
-      .porcentajeFinanciacion(50)
+      .porcentajeFinanciacion(BigDecimal.valueOf(50))
       .ajena(false)
       .build();
     //@formatter:on
